Changed watchdog interval

+++ web/webcontrol/sensors/power_meters.py
@@ -17,7 +17,7 @@ class LUMEL_N27P():
         corr_data = raw_data[0] | raw_data[1]
         corr_data = struct.unpack('!f', bytes.fromhex('{:x}'.format(corr_data)))[0]
 
-        return float('%.4f' % corr_data)
+        return int('%.4f' % corr_data)
 
     # def write(self, register_addr, data):
     #     return self.socket.write_float(registeraddress=register_addr)
